A single onion-service integration test
Adding to the py.test integration tests, this:

 - installs Chutney
 - uses it to build a local Tor test-network
 - sets up two storage servers on this test-network
 - proves that one can add a file, and the other can download it
